# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API-Person Finance Tools",
    "description": "Essa API tem como objetivo disponibilizar rotas para gestao de finanças",
    "termsOfService": "http://localhost:3000/terms",
    "contact": {
      "email": "fernandofv1110@gmail.com"
    },
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "API-BETA Personal Finance Tools"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Faça login na sua conta",
        "description": "entre na sua conta para receber seu token de acesso",
        "tags": ["auth"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              },
              "examples": {
                "product": {
                  "value": {
                    "email": "fe@gmail.com", 
                    "password": "123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Login"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Cadastre uma nova conta",
        "description": "Registre uma nova conta para poder logar na conta e ter seu token de acesso",
        "tags": ["auth"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              },
              "examples": {
                "product": {
                  "value": {
                    "name": "fe", 
                    "email": "fer@gmail.com", 
                    "password": "123"	, 
                    "confirmpassword": "123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "422": {
            "description": "Unprocessable Entity"
          },
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/accounting": {
      "post": {
        "summary": "Cadastro de recebimento/pagamentos",
        "description": "Essa rota será responsável por cadastrar um recebimento(earn) ou uma retirada(spend)",
        "tags": ["accounting"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "examples": {
                "product": {
                  "value": {
                    "name": "salario",
                    "earnOrSpend": "earn",
                    "price": 2200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Product Already exists!"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/accounting/all": {
      "get": {
        "description": "Busca o extrato da conta",
        "summary": "Lista com os itens da conta referente a ganho e gastos",
        "tags": ["accounting"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/accounting/summary": {
      "get": {
        "description": "resumo sobre a conta",
        "summary": " ira trazer o total de ganhos, total de gastos e o saldo atual  ",
        "tags": ["accounting"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      }
    },
    "/accounting/{id}": {
      "get": {
        "description": "Busca ganho ou retirada pelo id",
        "summary": "Busca uma informacao de despesa ou lucro pelo id",
        "tags": ["accounting"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do item",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Alterar um produto pelo id",
        "summary": "Alterar um produto pelo id",
        "tags": ["accounting"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Nome do produto pelo ID",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "examples": {
                "product": {
                  "value": {
                    "name": "salary",
                    "earnOrSpend": "earn",
                    "price": 4200
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Deletar um produto pelo id",
        "summary": "Deletar um produto pelo id",
        "tags": ["accounting"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id do produto",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "404": {
            "description": "NOT FOUND"
          }
        }
      }
    },
    "/finance/cdi": {
      "get": {
        "description": "Busca CDI atual",
        "summary": "exibe o valor do CDI atual",
        "tags": ["finance"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "data": {
                            "type": "string",
                            "format": "date"
                          },
                          "valor": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/finance/ipca": {
      "get": {
        "description": "Busca IPCA atual",
        "summary": "exibe o valor do IPCA atual",
        "tags": ["finance"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "data": {
                            "type": "string",
                            "format": "date"
                          },
                          "valor": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/finance/selic": {
      "get": {
        "description": "Busca SELIC atual",
        "summary": "exibe o valor do SELIC atual",
        "tags": ["finance"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "data": {
                            "type": "string",
                            "format": "date"
                          },
                          "valor": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/finance/stocks": {
      "get": {
        "description": "Busca SELIC atual",
        "summary": "exibe o valor do SELIC atual",
        "tags": ["finance"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "coluna": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "coluna": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "dados": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "Papel": {
                            "type": "string"
                          },
                          "Cotação": {
                            "type": "string"
                          },
                          "PsobreL": {
                            "type": "string"
                          },
                          "PsobreVP": {
                            "type": "string"
                          },
                          "PSR": {
                            "type": "string"
                          },
                          "DivYield": {
                            "type": "string"
                          },
                          "PsobreAtivo": {
                            "type": "string"
                          },
                          "PsobreCapGiro": {
                            "type": "string"
                          },
                          "PsobreEBIT": {
                            "type": "string"
                          },
                          "PsobreAtivCircLiq": {
                            "type": "string"
                          },
                          "EVsobreEBIT": {
                            "type": "string"
                          },
                          "EVsobreEBITDA": {
                            "type": "string"
                          },
                          "MrgEbit": {
                            "type": "string"
                          },
                          "MrgLíq": {
                            "type": "string"
                          },
                          "LiqCorr": {
                            "type": "string"
                          },
                          "ROIC": {
                            "type": "string"
                          },
                          "ROE": {
                            "type": "string"
                          },
                          "Liq2meses": {
                            "type": "string"
                          },
                          "PatrimLíq": {
                            "type": "string"
                          },
                          "DívBrutsobrePatrim": {
                            "type": "string"
                          },
                          "CrescRec5a": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "earnOrSpend": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "id": {
            "type": "string"
          }
        }
      },
      "Summary": {
        "type": "object",
        "properties": {
          "totalEarn": {
            "type": "number"
          },
          "totalSpend": {
            "type": "number"
          },
          "price": {
            "type": "number"
          }
        }
      },
      "Login": {
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "Register": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmpassword": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}